refactor: dto/in/ReportPostData business 폴더 하위로 이동
